Read Comment Article

+++ package/gitblog/src/database/migrations/2020_06_24_145900_create_post_views_table.php
@@ -13,10 +13,11 @@ class CreatePostViewsTable extends Migration
      */
     public function up()
     {
+
         Schema::create('post_views', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('post_id')->unsigned();
-            $table->integer('views');
+            $table->bigInteger('views');
             $table->timestamps();
         });
        Schema::table('post_views', function(Blueprint $table){
